# Performance testing of core code 'in isolation'