… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/Passes.h
… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/Passes.td
… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/Utils/AMDAIEUtils.cpp
… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/Utils/AMDAIEUtils.h
… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/test/CMakeLists.txt
… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/test/linalg_function_outlining_call_replication.mlir
… compiler/plugins/target/AMD-AIE/iree-amd-aie/Transforms/test/linalg_function_outlining_to_empty.mlir