To check the performance of the data movement, we added the flag 'replaceOutlinedFunctionsWithEmpty' which when used replaces calls to all compute (matmul/elementwise) functions with calls a function that does no compute. This means all the wall time reported must be spent moving data to and from the core from host memory. We currently don't have a way of answering the related question: what is the performance of the core code? In other words, if data movement was infinitely fast, how long would our matmul take to run? 

This PR makes it possible to answer that question. 

It's not obvious to me how we can eliminate _all_ data movement. There is probably a way to do this in isolation, from the input.o or input.asm files that peano generates. Maybe something like is [here](https://github.com/nod-ai/iree-amd-aie/pull/732/files). 

This PR gives the user the option, when a function is outlined, to wrap the new call site in a loop of user-chosen size. So if the loop count is n=100, the peano/ukernel code will run 100x, but without any more L3->L2->L1 data movement. In the limit n->infinity we get the core code performance

This PR subsumes the `replacOutlinedFunctionsWithEmpty' logic. 

The benchmark added in this PR would take  100'000 us if running at 100 %. With O3 it takes 3.5x that (see below). With O2 it takes 4.2x that (see below).  i.e. there is lots of room for improvement. 

```
matmul_512_512_512_bf16_f32_O2_npu1_4col_callrepl_100_outline_benchmark
--------------------------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------
BM_matmul/process_time/real_time_mean       424285 us         1996 us            3 items_per_second=2.35691/s
BM_matmul/process_time/real_time_median     424356 us         1992 us            3 items_per_second=2.35651/s
BM_matmul/process_time/real_time_stddev        147 us          108 us            3 items_per_second=818.927u/s
--------------------------------------------------------------------------------------------------
The largest program memory size (read from byte 72 of elf files) is 6096 bytes



matmul_512_512_512_bf16_f32_O3_npu1_4col_callrepl_100_outline_benchmark
--------------------------------------------------------------------------------------------------
Benchmark                                        Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------------
BM_matmul/process_time/real_time_mean       354238 us         1843 us            3 items_per_second=2.82297/s
BM_matmul/process_time/real_time_median     354047 us         1865 us            3 items_per_second=2.82449/s
BM_matmul/process_time/real_time_stddev        835 us         72.6 us            3 items_per_second=6.64549m/s
--------------------------------------------------------------------------------------------------
The largest program memory size (read from byte 72 of elf files) is 9568 bytes
```