@newling 2025-03-05
>  So if 100 calls run 354238 us, it means 3542.38 us on average for 1 call?

That is correct. 3542.38 us to do a 512x512x512 matmul on a single core. 

>  It seems much worse than a regular test result without 100 calls?

I'm not sure what you mean here. Let me explain the number from first principles 

The test runs a 512x512x512 matmul on a single AIE core, 100 times. 

That is 512*512*512*2 flops per call, so 268'435'456 flops per call. 

So that's a total of 26'843'545'600 flops for the 100 calls. 

It does this in  354238  us. 

That is  0.354238 seconds (1'000'000 microseconds in a second). 

So it's chugging out  26'843'545'600 /  0.354238 per second. 

That's 75778277880.97269 flops per second, which is 

 0.07577827788097269 teraflops per second (divide by 1e12). 
 
 That's one core. We know that phoenix can do 4 tflops with 16 cores, so the theoretical peak for 1 core is 4/16 = 0.25 tflops. 
 
Dividing observed tflops by theoretical peak tflops,  this test shows we're getting 

0.07577827788097269 / 0.25 = 0.303  of peak theoretical performance. That is 30.3 % of theoretical peak