@yzhang93 2025-03-05
> > So if 100 calls run 354238 us, it means 3542.38 us on average for 1 call?
> 
> That is correct. 3542.38 us to do a 512x512x512 matmul on a single core.
> 
> > It seems much worse than a regular test result without 100 calls?
> 
> I'm not sure what you mean here. Let me explain the number from first principles

Thanks for the explanation, I didn't realize it's on a single core.