@konstantinschwarz 2025-03-05
Drive-by comment: have you tried using performance counters (events) to measure the performance of the compute code?