@Yu-Zhewen 2025-03-06
> > Drive-by comment: have you tried using performance counters (events) to measure the performance of the compute code?
> 
> Not really. Do you mean something AIE/npu specific?

Yes, it's an AIE feature that enables cycle-accurate performance measurement using timers within the cores.
https://github.com/Xilinx/mlir-aie/tree/main/programming_guide/section-4/section-4b

To support this, it requires some routing resources and also sets up special registers. There is also existing code in `mlir-air` to automate this process:

https://github.com/Xilinx/mlir-air/blob/7301aa0b6be83806554e5f739c55a37e6bb87d56/mlir/lib/Conversion/AIRToAIEPass.cpp#L3189

https://github.com/Xilinx/mlir-air/blob/7301aa0b6be83806554e5f739c55a37e6bb87d56/mlir/lib/Conversion/AIRRtToNpuPass.cpp#L1324

The only issue is that setting up such performance tracing for a full-array design can be challenging due to routing congestion (unless the design itself also uses packet flows). However, it should work fine for a single core, a single column, or 2×2 cases.