@newling 2025-03-06
Thanks @Yu-Zhewen I'll take a look. It'd be better to have a cycle accurate evaluation.